refactor: 사용되지 않는 dto 삭제, DIP 원칙에 의한 일부 파일 위치 수정
- 모든 페이로드를 application.event ->  infrastructure.messaging.payload로 이동
- ScheduleCreatedPayload domain event로 이동
- 일부 사용하지 않는 dto 삭제
- 참조 서비스 패키지 위치 변경
