+++ src/main/java/com/firstticket/programservice/domain/event/ScheduleCreatedEventData.java
@@ -1,4 +1,4 @@
-package com.firstticket.programservice.application.event;
+package com.firstticket.programservice.domain.event;
 
 import java.util.List;
 import java.util.UUID;
@@ -6,10 +6,12 @@ import java.util.UUID;
 import com.firstticket.programservice.domain.service.dto.VenueValidationData;
 
 /**
- * 스케줄 생성 이벤트 페이로드 (토픽: schedule.created).
- * 좌석 서비스가 수신하여 BookingSeat을 생성한다.
+ * schedule.created 이벤트의 seatTemplates 구성 데이터.
+ * 도메인 계층에서 사용하는 이벤트 데이터 DTO.
+ * infrastructure/event/ProgramEventPublisherImpl에서
+ * ScheduleCreatedPayload.SeatTemplate으로 변환된다.
  */
-public record ScheduleCreatedPayload(
+public record ScheduleCreatedEventData(
     UUID scheduleId,
     UUID programId,
     List<SeatTemplate> seatTemplates
@@ -17,8 +19,7 @@ public record ScheduleCreatedPayload(
 
     /**
      * 구역별 좌석 템플릿.
-     * SEATED  : rowCount × colCount 개 BookingSeat 생성
-     * STANDING·FREE: capacity 기반 재고 관리
+     * VenueValidationData.SectionInfo + price로 구성된다.
      */
     public record SeatTemplate(
         UUID sectionId,
