--- src/main/java/com/firstticket/programservice/infrastructure/client/dto/SectionInfoResponse.java
@@ -1,9 +1,0 @@
-package com.firstticket.programservice.infrastructure.client.dto;
-
-public record SectionInfoResponse(
-    String sectionName,  // ← 추가
-    Integer rowCount,
-    Integer colCount,
-    Integer capacity
-) {
-}
