+++ src/main/java/com/firstticket/programservice/infrastructure/messaging/payload/ProgramCreatedPayload.java
@@ -1,4 +1,6 @@
-package com.firstticket.programservice.application.event;
+// infrastructure/messaging/payload/ProgramCreatedPayload.java
+
+package com.firstticket.programservice.infrastructure.messaging.payload;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -7,14 +9,9 @@ import com.firstticket.programservice.domain.Program;
 import com.firstticket.programservice.domain.ProgramStatus;
 
 /**
- * 프로그램 생성 이벤트 페이로드 (토픽: program.created).
+ * program.created 토픽 Kafka 페이로드.
  * 대기열 서비스가 수신하여 프로그램 대기열을 초기화한다.
- * createProgram() 시 발행한다.
- *
- * status는 항상 DRAFT로 고정된다.
- * openAt: 가장 빠른 스케줄의 saleStartAt
- * closeAt: 가장 늦은 스케줄의 saleEndAt
- * → 스케줄이 없는 DRAFT 초기 생성 시점이므로 null 허용
+ * 생성 시점에는 스케줄이 없으므로 openAt·closeAt은 null로 발행한다.
  */
 public record ProgramCreatedPayload(
     UUID programId,
