+++ src/main/java/com/firstticket/programservice/infrastructure/messaging/payload/ProgramTimeUpdatedPayload.java
@@ -1,4 +1,4 @@
-package com.firstticket.programservice.application.event;
+package com.firstticket.programservice.infrastructure.messaging.payload;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -7,15 +7,9 @@ import com.firstticket.programservice.domain.Program;
 import com.firstticket.programservice.domain.Schedule;
 
 /**
- * 프로그램 판매 기간 변경 이벤트 페이로드 (토픽: program.time.updated).
+ * program.time.updated 토픽 Kafka 페이로드.
  * 대기열 서비스가 수신하여 openAt·closeAt을 갱신한다.
- *
- * 발행 시점:
- * - publishProgram(): ON_SALE 전환 시
- * - updateSchedule(): 판매 기간 변경 시
- *
- * openAt: 변경된 saleStartAt
- * closeAt: 변경된 saleEndAt
+ * publishProgram() / updateSchedule() 에서 발행한다.
  */
 public record ProgramTimeUpdatedPayload(
     UUID programId,
